+++ chainkit/src/lib/config.ts
@@ -82,8 +82,8 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "1.0.0",
-  sdkVersion: "0.3.0-alpha.11",
-  genVersion: "2.755.9",
+  sdkVersion: "0.3.0-alpha.12",
+  genVersion: "2.774.0",
   userAgent:
-    "speakeasy-sdk/typescript 0.3.0-alpha.11 2.755.9 1.0.0 @avalanche-sdk/chainkit",
+    "speakeasy-sdk/typescript 0.3.0-alpha.12 2.774.0 1.0.0 @avalanche-sdk/chainkit",
 } as const;