+++ chainkit/src/models/components/erc1155transfer.ts
@@ -25,9 +25,13 @@ export type Erc1155Transfer = {
    */
   blockNumber: string;
   /**
-   * The block finality timestamp.
+   * The block creation (proposal) timestamp in seconds
    */
   blockTimestamp: number;
+  /**
+   * The block creation (proposal) timestamp in milliseconds. Available only after Granite upgrade.
+   */
+  blockTimestampMilliseconds?: number | undefined;
   /**
    * The block hash identifier.
    */
@@ -51,6 +55,7 @@ export const Erc1155Transfer$inboundSchema: z.ZodType<
 > = z.object({
   blockNumber: z.string(),
   blockTimestamp: z.number(),
+  blockTimestampMilliseconds: z.number().optional(),
   blockHash: z.string(),
   txHash: z.string(),
   from: RichAddress$inboundSchema,
@@ -63,6 +68,7 @@ export const Erc1155Transfer$inboundSchema: z.ZodType<
 export type Erc1155Transfer$Outbound = {
   blockNumber: string;
   blockTimestamp: number;
+  blockTimestampMilliseconds?: number | undefined;
   blockHash: string;
   txHash: string;
   from: RichAddress$Outbound;
@@ -80,6 +86,7 @@ export const Erc1155Transfer$outboundSchema: z.ZodType<
 > = z.object({
   blockNumber: z.string(),
   blockTimestamp: z.number(),
+  blockTimestampMilliseconds: z.number().optional(),
   blockHash: z.string(),
   txHash: z.string(),
   from: RichAddress$outboundSchema,