+++ chainkit/src/models/operations/posttransactionexportjob.ts
@@ -14,7 +14,15 @@ export const PostTransactionExportJobServerList = [
 
 export type PostTransactionExportJobRequest =
   | components.CreateEvmTransactionExportRequest
-  | components.CreatePrimaryNetworkTransactionExportRequest;
+  | (components.CreatePrimaryNetworkTransactionExportRequest & {
+    type: "TRANSACTION_EXPORT_PRIMARY_NETWORK";
+  })
+  | (components.CreatePrimaryNetworkTransactionExportRequest & {
+    type: "TRANSACTION_EXPORT_PRIMARY_NETWORK_STAKING";
+  })
+  | (components.CreatePrimaryNetworkTransactionExportRequest & {
+    type: "TRANSACTION_EXPORT_PRIMARY_NETWORK_SIMPLE";
+  });
 
 /** @internal */
 export const PostTransactionExportJobRequest$inboundSchema: z.ZodType<
@@ -23,12 +31,28 @@ export const PostTransactionExportJobRequest$inboundSchema: z.ZodType<
   unknown
 > = z.union([
   components.CreateEvmTransactionExportRequest$inboundSchema,
-  components.CreatePrimaryNetworkTransactionExportRequest$inboundSchema,
+  components.CreatePrimaryNetworkTransactionExportRequest$inboundSchema.and(
+    z.object({ type: z.literal("TRANSACTION_EXPORT_PRIMARY_NETWORK") }),
+  ),
+  components.CreatePrimaryNetworkTransactionExportRequest$inboundSchema.and(
+    z.object({ type: z.literal("TRANSACTION_EXPORT_PRIMARY_NETWORK_STAKING") }),
+  ),
+  components.CreatePrimaryNetworkTransactionExportRequest$inboundSchema.and(
+    z.object({ type: z.literal("TRANSACTION_EXPORT_PRIMARY_NETWORK_SIMPLE") }),
+  ),
 ]);
 /** @internal */
 export type PostTransactionExportJobRequest$Outbound =
   | components.CreateEvmTransactionExportRequest$Outbound
-  | components.CreatePrimaryNetworkTransactionExportRequest$Outbound;
+  | (components.CreatePrimaryNetworkTransactionExportRequest$Outbound & {
+    type: "TRANSACTION_EXPORT_PRIMARY_NETWORK";
+  })
+  | (components.CreatePrimaryNetworkTransactionExportRequest$Outbound & {
+    type: "TRANSACTION_EXPORT_PRIMARY_NETWORK_STAKING";
+  })
+  | (components.CreatePrimaryNetworkTransactionExportRequest$Outbound & {
+    type: "TRANSACTION_EXPORT_PRIMARY_NETWORK_SIMPLE";
+  });
 
 /** @internal */
 export const PostTransactionExportJobRequest$outboundSchema: z.ZodType<
@@ -37,7 +61,15 @@ export const PostTransactionExportJobRequest$outboundSchema: z.ZodType<
   PostTransactionExportJobRequest
 > = z.union([
   components.CreateEvmTransactionExportRequest$outboundSchema,
-  components.CreatePrimaryNetworkTransactionExportRequest$outboundSchema,
+  components.CreatePrimaryNetworkTransactionExportRequest$outboundSchema.and(
+    z.object({ type: z.literal("TRANSACTION_EXPORT_PRIMARY_NETWORK") }),
+  ),
+  components.CreatePrimaryNetworkTransactionExportRequest$outboundSchema.and(
+    z.object({ type: z.literal("TRANSACTION_EXPORT_PRIMARY_NETWORK_STAKING") }),
+  ),
+  components.CreatePrimaryNetworkTransactionExportRequest$outboundSchema.and(
+    z.object({ type: z.literal("TRANSACTION_EXPORT_PRIMARY_NETWORK_SIMPLE") }),
+  ),
 ]);
 
 export function postTransactionExportJobRequestToJSON(