+++ chainkit/src/sdk/evmtransactions.ts
@@ -16,7 +16,7 @@ export class EvmTransactions extends ClientSDK {
    * List the latest transactions across all supported EVM chains
    *
    * @remarks
-   * Lists the most recent transactions from all supported EVM-compatible  chains. The results can be filtered based on transaction status.
+   * Lists the most recent transactions from all supported EVM-compatible chains. The results can be filtered based on transaction status.
    */
   async listLatestAllChains(
     request: operations.ListLatestTransactionsAllChainsRequest,